02_document_intelligence_extracao/requirements.txt
# Dependências do projeto
# Instalar com: pip install -r requirements.txt

fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.1.0
requests==2.32.3
